Refactor test_protocol.py
